# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "B7Store API",
        "description": "API completa para o sistema B7Store - E-commerce",
        "contact": {
            "email": "suporte@b7store.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "https://api.b7store.com",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registrar novo usuário",
                "description": "Cria um novo usuário no sistema",
                "operationId": "d03ee3a44bdd6f0025efd14c5c76b5ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    },
                    "500": {
                        "description": "Erro ao criar usuário"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login de usuário",
                "description": "Autentica usuário e retorna token Bearer",
                "operationId": "9f6f4aca790a5229fb638e798623c6e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/user/addresses": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Listar endereços",
                "description": "Retorna todos os endereços do usuário autenticado",
                "operationId": "5538c6eded480ef3d257133a81eb1076",
                "responses": {
                    "200": {
                        "description": "Lista de endereços"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Adicionar endereço",
                "description": "Cadastra um novo endereço para o usuário autenticado",
                "operationId": "434614f3c629c07413f55f571964902e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "zipcode",
                                    "street",
                                    "city",
                                    "state",
                                    "country"
                                ],
                                "properties": {
                                    "zipcode": {
                                        "type": "string",
                                        "example": "12345-678"
                                    },
                                    "street": {
                                        "type": "string",
                                        "example": "Rua das Flores, 123"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "São Paulo"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "SP"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Brasil"
                                    },
                                    "complement": {
                                        "type": "string",
                                        "example": "Apto 42"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço cadastrado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticação via Laravel Sanctum. Use: Bearer {seu-token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}